[fix] : Refactor to remove name conflict
