+++ run.py
@@ -2,6 +2,6 @@
     Serve the api locally
 """
 
-from api.api import API
+from api.api_v1 import API
 if __name__ == '__main__':
     API.run()
